requirements.txt
Flask==2.3.2
torch==2.5.0
torchvision==0.20.0
Pillow==9.5.0
flask-cors
numpy
scipy
werkzeug